+++ components/clarinet-cli/src/frontend/cli.rs
@@ -32,6 +32,7 @@ use clarity_repl::clarity::vm::analysis::AnalysisDatabase;
 use clarity_repl::clarity::vm::costs::LimitedCostTracker;
 use clarity_repl::clarity::vm::types::QualifiedContractIdentifier;
 use clarity_repl::clarity::ClarityVersion;
+use clarity_repl::frontend::terminal::print_clarity_wasm_warning;
 use clarity_repl::repl::diagnostic::output_diagnostic;
 use clarity_repl::repl::{ClarityCodeSource, ClarityContract, ContractDeployer, DEFAULT_EPOCH};
 use clarity_repl::{analysis, repl, Terminal};
@@ -347,6 +348,9 @@ struct Console {
         conflicts_with = "use_on_disk_deployment_plan"
     )]
     pub use_computed_deployment_plan: bool,
+    /// Allow the Clarity Wasm preview to run in parallel with the Clarity interpreter (beta)
+    #[clap(long = "enable-clarity-wasm")]
+    pub enable_clarity_wasm: bool,
 }
 
 #[derive(Parser, PartialEq, Clone, Debug)]
@@ -929,9 +933,50 @@ pub fn main() {
                             std::process::exit(1);
                         }
 
-                        Terminal::load(artifacts.session)
+                        if cmd.enable_clarity_wasm {
+                            let mut manifest_wasm = manifest.clone();
+                            manifest_wasm.repl_settings.clarity_wasm_mode = true;
+                            let (_, _, artifacts_wasm) = load_deployment_and_artifacts_or_exit(
+                                &manifest_wasm,
+                                &cmd.deployment_plan_path,
+                                cmd.use_on_disk_deployment_plan,
+                                cmd.use_computed_deployment_plan,
+                            );
+                            if artifacts.success != artifacts_wasm.success {
+                                for contract in deployment.contracts.keys() {
+                                    let empty_diag = vec![];
+                                    let diag = artifacts
+                                        .diags
+                                        .get(contract)
+                                        .unwrap_or(empty_diag.as_ref());
+                                    let diag_wasm = artifacts_wasm
+                                        .diags
+                                        .get(contract)
+                                        .unwrap_or(empty_diag.as_ref());
+
+                                    if diag.len() != diag_wasm.len() {
+                                        dbg!(&diag);
+                                        dbg!(&diag_wasm);
+                                    }
+                                }
+                                print_clarity_wasm_warning();
+                            }
+
+                            Terminal::load(artifacts.session, Some(artifacts_wasm.session))
+                        } else {
+                            Terminal::load(artifacts.session, None)
+                        }
                     }
-                    None => Terminal::new(repl::SessionSettings::default()),
+                    None => {
+                        let settings = repl::SessionSettings::default();
+                        if cmd.enable_clarity_wasm {
+                            let mut settings_wasm = repl::SessionSettings::default();
+                            settings_wasm.repl_settings.clarity_wasm_mode = true;
+                            Terminal::new(settings, Some(settings_wasm))
+                        } else {
+                            Terminal::new(settings, None)
+                        }
+                    }
                 };
                 let reload = terminal.start();