+++ components/clarinet-cli/src/generate/project.rs
@@ -415,14 +415,14 @@ disable_stacks_api = false
 
 # Send some stacking orders
 [[devnet.pox_stacking_orders]]
-start_at_cycle = 0
+start_at_cycle = 2
 duration = 12
 wallet = "wallet_1"
 slots = 2
 btc_address = "mr1iPkD9N3RJZZxXRk7xF9d36gffa6exNC"
 
 [[devnet.pox_stacking_orders]]
-start_at_cycle = 1
+start_at_cycle = 2
 duration = 12
 wallet = "wallet_2"
 slots = 1